src/requirements_hpc.txt
wheel
einops
matplotlib
numpy<2
torch
pandas
Pillow
scikit_image
scikit-learn
scipy
setuptools
SimpleITK
timm
ninja
torch_summary
torchsummary
torchvision
tensorboard
torchio
monai
psutil
lightning
tifffile